mcps/crm_mcp/requirements.txt
alembic
sqlalchemy
uvicorn
fastapi
pydantic
psycopg2-binary
pgvector
mcp[stdio]
google-auth
requests
